V1.1.1 - Spring Security - Cambios a CORS

+++ src/main/java/com/iosepbp/bienes_patrimoniales/Modelos/Bien.java
@@ -89,7 +89,7 @@ public class Bien {
         this.id_bien = id_bien;
     }
     public Integer getId_bien(){
-        return this.id_bien;
+        return (Integer) this.id_bien;
     }
 
     //cod_bien
